requirements.txt
streamlit==1.52.2
langchain==1.2.0
langchain-google-genai==4.1.2
langchain-community==0.4.1
langchain-core==1.2.6
langchain-chroma==1.1.0
langchain_text_splitters==1.1.0
chromadb==1.4.0
pypdf==6.5.0
python-dotenv==1.2.1